requirements.txt
aiogram-utils==0.1.4
aiohttp==3.8.6
Flask==3.0.3